# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zenin — Android device monitoring dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Step 1 — verify credentials, send OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Credentials valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginStepResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/otp/verify:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Step 2 — verify OTP, receive session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyInput"
      responses:
        "200":
          description: OTP valid, token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "401":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /auth/sessions:
    get:
      operationId: getSessions
      tags: [auth]
      summary: List active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionsResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Terminate all active sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [auth]
      summary: Revoke a specific session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ── Panel ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: ["string", "null"]
      required: [ok]

    # ── Auth ────────────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [username, userId, password]
      properties:
        username:
          type: string
        userId:
          type: string
        password:
          type: string
    LoginStepResult:
      type: object
      required: [step]
      properties:
        step:
          type: string
          enum: [otp, done]
        preAuthId:
          type: ["string", "null"]
        token:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/UserBrief"
    OtpVerifyInput:
      type: object
      required: [preAuthId, code]
      properties:
        preAuthId:
          type: string
        code:
          type: string
    AuthSuccess:
      type: object
      required: [step, token, user]
      properties:
        step:
          type: string
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserBrief"
    UserBrief:
      type: object
      required: [id, userId, username, name, role]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
    UserProfile:
      type: object
      required: [id, userId, username, name, role, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        name:
          type: string
        role:
          type: string
        tgUsername:
          type: ["string", "null"]
        accessExpiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    SessionsResponse:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/Session"
    Session:
      type: object
      required: [id, userId, createdAt, lastActiveAt]
      properties:
        id:
          type: string
        userId:
          type: string
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: string
        lastActiveAt:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string

    # ── Panel ───────────────────────────────────────────────────────────────
...